requirements.txt
openai
streamlit
phidata
googlesearch-python
pycountry
hmac